Revert "[NoQA] Tests for group chat name"
