+++ tests/ui/UnreadIndicatorsTest.tsx
@@ -4,26 +4,30 @@ import {act, fireEvent, render, screen, waitFor} from '@testing-library/react-na
 import {addSeconds, format, subMinutes, subSeconds} from 'date-fns';
 import {utcToZonedTime} from 'date-fns-tz';
 import React from 'react';
-import {AppState, DeviceEventEmitter} from 'react-native';
+import {AppState, DeviceEventEmitter, Linking} from 'react-native';
 import type {ViewStyle} from 'react-native';
 import type {OnyxEntry} from 'react-native-onyx';
 import Onyx from 'react-native-onyx';
+import type Animated from 'react-native-reanimated';
 import * as CollectionUtils from '@libs/CollectionUtils';
 import DateUtils from '@libs/DateUtils';
 import * as Localize from '@libs/Localize';
 import LocalNotification from '@libs/Notification/LocalNotification';
 import * as NumberUtils from '@libs/NumberUtils';
+import * as Pusher from '@libs/Pusher/pusher';
+import PusherConnectionManager from '@libs/PusherConnectionManager';
 import FontUtils from '@styles/utils/FontUtils';
 import * as AppActions from '@userActions/App';
 import * as Report from '@userActions/Report';
 import * as User from '@userActions/User';
 import App from '@src/App';
+import CONFIG from '@src/CONFIG';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
+import appSetup from '@src/setup';
 import type {ReportAction, ReportActions} from '@src/types/onyx';
 import PusherHelper from '../utils/PusherHelper';
 import * as TestHelper from '../utils/TestHelper';
-import {navigateToSidebarOption} from '../utils/TestHelper';
 import waitForBatchedUpdates from '../utils/waitForBatchedUpdates';
 import waitForBatchedUpdatesWithAct from '../utils/waitForBatchedUpdatesWithAct';
 
@@ -41,6 +45,12 @@ jest.mock('react-native/Libraries/LogBox/LogBox', () => ({
         ignoreLogs: jest.fn(),
         ignoreAllLogs: jest.fn(),
     },
+}));
+
+jest.mock('react-native-reanimated', () => ({
+    ...jest.requireActual<typeof Animated>('react-native-reanimated/mock'),
+    createAnimatedPropAdapter: jest.fn,
+    useReducedMotion: jest.fn,
 }));
 
 /**
@@ -105,7 +115,23 @@ jest.mock('@react-navigation/native', () => {
 });
 
 beforeAll(() => {
-    TestHelper.beforeAllSetupUITests(true);
+    // In this test, we are generically mocking the responses of all API requests by mocking fetch() and having it
+    // return 200. In other tests, we might mock HttpUtils.xhr() with a more specific mock data response (which means
+    // fetch() never gets called so it does not need mocking) or we might have fetch throw an error to test error handling
+    // behavior. But here we just want to treat all API requests as a generic "success" and in the cases where we need to
+    // simulate data arriving we will just set it into Onyx directly with Onyx.merge() or Onyx.set() etc.
+    global.fetch = TestHelper.getGlobalFetchMock();
+
+    Linking.setInitialURL('https://new.expensify.com/');
+    appSetup();
+
+    // Connect to Pusher
+    PusherConnectionManager.init();
+    Pusher.init({
+        appKey: CONFIG.PUSHER.APP_KEY,
+        cluster: CONFIG.PUSHER.CLUSTER,
+        authEndpoint: `${CONFIG.EXPENSIFY.DEFAULT_API_ROOT}api/AuthenticatePusher?`,
+    });
 });
 
 function scrollUpToRevealNewMessagesBadge() {
@@ -145,6 +171,13 @@ function navigateToSidebar(): Promise<void> {
         fireEvent(reportHeaderBackButton, 'press');
     }
     return waitForBatchedUpdates();
+}
+
+async function navigateToSidebarOption(index: number): Promise<void> {
+    const hintText = Localize.translateLocal('accessibilityHints.navigatesToChat');
+    const optionRows = screen.queryAllByAccessibilityHint(hintText);
+    fireEvent(optionRows[index], 'press');
+    await waitForBatchedUpdatesWithAct();
 }
 
 function areYouOnChatListScreen(): boolean {
